Provide needed changes to successfully publish component on npm
This commit removes eslint-loader from building app on dev and prod
environment. Eslint-loader blocked compiling because of minor eslint
issues. Options like 'emitError: false' and 'emitWarining: true' for
eslint-loader didn't fix this problem and compiling on prod
couldn't finish.

For this reason eslint-loader have been removed but finding the way to
load code with eslint-loader and using proper options not preventing
it from compiling can be treated as TO DO.